# Conflicts:
#	src/model/MoteurImpl.java

+++ src/vue/Curseur.java
@@ -6,6 +6,6 @@ import command.Command;
  * Created by tp15009314 on 16/11/16.
  */
 public interface Curseur {
-    public int getPosition();
+    public double getPosition();
     public void setCommand(Command c);
 }
